Refactored OptionsModalList component
