+++ app/screens/options_modal/options_modal_list.tsx
@@ -2,7 +2,7 @@
 // See LICENSE.txt for license information.
 
 import React, {useCallback} from 'react';
-import {StyleSheet, TouchableOpacity, View} from 'react-native';
+import {Platform, StyleSheet, TouchableOpacity, View} from 'react-native';
 
 import FormattedText from '@components/formatted_text';
 import OptionsItem from '@components/options_item';
@@ -29,6 +29,20 @@ const OptionsModalList = ({items = [], onCancelPress, onItemPress, title}: Props
             }
         }, 100);
     }), []);
+
+    if (Platform.OS === 'android') {
+        return (
+            <View style={style.wrapper}>
+                <View style={style.optionContainer}>
+                    <OptionsItem
+                        onHandleItemPress={handleItemPress}
+                        onHandleCancelPress={handleCancelPress}
+                        items={items}
+                    />
+                </View>
+            </View>
+        );
+    }
 
     return (
         <View style={style.container}>
@@ -65,7 +79,11 @@ const style = StyleSheet.create({
         flexDirection: 'row',
         justifyContent: 'space-between',
         padding: 15,
-        width: '100%',
+        ...Platform.select({
+            ios: {
+                width: '100%',
+            },
+        }),
     },
     optionBorder: {
         borderBottomWidth: 1,
@@ -80,17 +98,42 @@ const style = StyleSheet.create({
     optionContainer: {
         alignSelf: 'stretch',
         backgroundColor: 'white',
-        borderRadius: 12,
-        marginBottom: 20,
-        marginHorizontal: 20,
+        ...Platform.select({
+            ios: {
+                borderRadius: 12,
+                marginBottom: 20,
+                marginHorizontal: 20,
+            },
+            android: {
+                borderRadius: 2,
+                marginHorizontal: 30,
+            },
+        }),
     },
     optionIcon: {
-        color: '#4E8ACC',
+        ...Platform.select({
+            ios: {
+                color: '#4E8ACC',
+            },
+            android: {
+                color: '#7f8180',
+            },
+        }),
     },
     optionText: {
-        color: '#4E8ACC',
-        flex: 1,
-        fontSize: 20,
+        ...Platform.select({
+            ios: {
+                color: '#4E8ACC',
+                flex: 1,
+                fontSize: 20,
+            },
+            android: {
+                color: '#000',
+                flex: 1,
+                fontSize: 16,
+            },
+        }),
+
     },
     optionTitleText: {
         color: '#7f8180',
@@ -103,8 +146,17 @@ const style = StyleSheet.create({
         justifyContent: 'flex-end',
     },
     wrapper: {
-        maxWidth: 450,
-        width: '100%',
+        ...Platform.select({
+            ios: {
+                maxWidth: 450,
+                width: '100%',
+            },
+            android: {
+                flex: 1,
+                alignItems: 'center',
+                justifyContent: 'center',
+            },
+        }),
     },
 });
